requirements.txt
configparser==3.5.0
msgpack==0.5.6
setuptools==28.8.1
kafka-python==1.4.2